implement live traces
